// types/profileLookup.ts
export type Profile = {
        player_id: number;
        "@id": string;
        url: string;
        username: string;
        followers: number;
        country: string;
        last_online: number;
        joined: number;
        status: string;
        is_streamer: boolean;
        verified: boolean;
    }

    export default Profile